Added Personal Drawing Register for Admin, Roles, Seed Data for States

+++ DrawingRegisterWeb/Controllers/ProjectStatesController.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using DrawingRegisterWeb.Data;
 using DrawingRegisterWeb.Models;
@@ -26,18 +21,20 @@ namespace DrawingRegisterWeb.Controllers
 		// GET: ProjectStates
 		public async Task<IActionResult> Index(string search, string states)
 		{
-			var claimsIdentity = (ClaimsIdentity)User.Identity;
+			var claimsIdentity = (ClaimsIdentity)User.Identity!;
 			var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
 
+			var drawingRegisters = from dr in _context.DrawingRegisters select dr;
+			var drawingRegister = drawingRegisters.FirstOrDefault(dr => dr.UserId == claim!.Value);
+
+			if (drawingRegister == null) return RedirectToAction("Index", "DrawingRegisters");
+
 			var projectStates = from p in _context.ProjectState select p;
-			var drawingRegisters = from dr in _context.drawingRegisters select dr;
-			var drawingRegister = drawingRegisters.FirstOrDefault(dr => dr.UserId == claim.Value);
 
 			if (search != null)
 			{
-				projectStates = projectStates.Where(p =>
-				p.Name.Contains(search) ||
-				p.Description.Contains(search));
+				projectStates = projectStates.Where(p => p.Name.Contains(search) || 
+					p.Description.Contains(search));
 			}
 
 			if (states != null)
@@ -63,30 +60,30 @@ namespace DrawingRegisterWeb.Controllers
 
 			var projectStateVM = new ProjectStateVM
 			{
-				ProjectStates = await projectStates.Where(p => p.DrawingRegisterId == drawingRegister.Id).OrderBy(p => p.Id).ToListAsync()
+				ProjectStates = await projectStates.Where(p => p.DrawingRegisterId == drawingRegister!.Id).OrderBy(p => p.Id).ToListAsync(),
+				Search = search,
+				States = states
 			};
 
 			return View(projectStateVM);
 		}
 
-		// GET: ProjectStates/Create
+		//ProjectStates/Create
 		public IActionResult Create()
 		{
-			var claimsIdentity = (ClaimsIdentity)User.Identity;
+			var claimsIdentity = (ClaimsIdentity)User.Identity!;
 			var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
-
-			var drawingRegisters = from dr in _context.drawingRegisters select dr;
-			var drawingRegister = drawingRegisters.FirstOrDefault(dr => dr.UserId == claim.Value);
+			var drawingRegisters = from dr in _context.DrawingRegisters select dr;
+			var drawingRegister = drawingRegisters.FirstOrDefault(dr => dr.UserId == claim!.Value);
 
 			ProjectState projectState = new()
 			{
-				DrawingRegisterId = drawingRegister.Id
+				DrawingRegisterId = drawingRegister!.Id
 			};
 
 			return View(projectState);
 		}
 
-		// POST: ProjectStates/Create
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> Create(ProjectState projectState)
@@ -100,7 +97,7 @@ namespace DrawingRegisterWeb.Controllers
 			return View(projectState);
 		}
 
-		// GET: ProjectStates/Edit
+		//ProjectStates/Edit
 		public async Task<IActionResult> Edit(int? id)
 		{
 			if (id == null || _context.ProjectState == null)
@@ -116,7 +113,6 @@ namespace DrawingRegisterWeb.Controllers
 			return View(projectState);
 		}
 
-		// POST: ProjectStates/Edit
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> Edit(int id, ProjectState projectState)
@@ -135,54 +131,38 @@ namespace DrawingRegisterWeb.Controllers
 				}
 				catch (DbUpdateConcurrencyException)
 				{
-					if (!ProjectStateExists(projectState.Id))
-					{
-						return NotFound();
-					}
-					else
-					{
-						throw;
-					}
+					if (!ProjectStateExists(projectState.Id)) return NotFound();
+					else throw;
 				}
 				return RedirectToAction(nameof(Index));
 			}
 			return View(projectState);
 		}
 
-		// GET: ProjectStates/Delete
+		//ProjectStates/Delete
 		public async Task<IActionResult> Delete(int? id)
 		{
-			if (id == null || _context.ProjectState == null)
-			{
-				return NotFound();
-			}
+			if (id == null || _context.ProjectState == null) return NotFound();
 
-			var projectState = await _context.ProjectState
-				.FirstOrDefaultAsync(m => m.Id == id);
-			if (projectState == null)
-			{
-				return NotFound();
-			}
+			var projectState = await _context.ProjectState.FirstOrDefaultAsync(m => m.Id == id);
+
+			if (projectState == null) return NotFound();
 
 			return View(projectState);
 		}
 
-		// POST: ProjectStates/Delete/5
 		[HttpPost, ActionName("Delete")]
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> DeleteConfirmed(int id)
 		{
-			if (_context.ProjectState == null)
-			{
-				return Problem("Entity set is null.");
-			}
+			if (_context.ProjectState == null) return Problem("Entity set is null.");
+
 			var projectState = await _context.ProjectState.FindAsync(id);
-			if (projectState != null)
-			{
-				_context.ProjectState.Remove(projectState);
-			}
+
+			if (projectState != null) _context.ProjectState.Remove(projectState);
 
 			await _context.SaveChangesAsync();
+
 			return RedirectToAction(nameof(Index));
 		}
